Moved saved output button and added two new icons in case they are needed
