+++ frontend/ohio-house-scraper-app/src/components/SaveOutputButton.module.css
@@ -1,0 +1,30 @@
+.mainContainer {
+    width: 100%;
+    height: 1.5rem;
+    padding: 0 0.5rem;
+
+    border: none;
+    border-radius: 0.15rem;
+    outline: 1px solid rgb(141, 141, 141);
+    background-color: rgb(41, 41, 54);
+    overflow: hidden;
+    cursor: pointer;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.mainContainer:hover {
+    transform: translateY(-1px);
+}
+
+.mainContainer:active {
+    transform: translateY(1px);
+}
+
+.mainContainer > p {
+    color: white;
+    font-family: "poppins";
+    font-size: 0.9rem;
+}
