+++ frontend/ohio-house-scraper-app/src/components/Toggle.js
@@ -1,0 +1,7 @@
+import classes from "./Toggle.module.css"
+
+const Toggle = () => {
+    return (<div className={classes.mainContainer}></div>)
+}
+
+export default Toggle
